refactoring spliting test data controller into multiple controllers
