+++ src/main/java/uk/gov/companieshouse/api/testdata/service/CombinedSicActivitiesService.java
@@ -1,0 +1,8 @@
+package uk.gov.companieshouse.api.testdata.service;
+
+import uk.gov.companieshouse.api.testdata.model.rest.request.CombinedSicActivitiesRequest;
+import uk.gov.companieshouse.api.testdata.model.rest.response.CombinedSicActivitiesResponse;
+
+public interface CombinedSicActivitiesService extends DataService<CombinedSicActivitiesResponse,
+        CombinedSicActivitiesRequest> {
+}
