+++ src/test/java/uk/gov/companieshouse/api/testdata/controller/AppealControllerTest.java
@@ -1,0 +1,73 @@
+package uk.gov.companieshouse.api.testdata.controller;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import uk.gov.companieshouse.api.testdata.model.rest.request.DeleteAppealsRequest;
+import uk.gov.companieshouse.api.testdata.service.AppealsService;
+
+@ExtendWith(MockitoExtension.class)
+class AppealControllerTest {
+    @Mock
+    private AppealsService appealsService;
+
+    @InjectMocks
+    private AppealController appealController;
+
+
+    @Test
+    void deleteAppealSuccess() {
+        DeleteAppealsRequest request = new DeleteAppealsRequest();
+        request.setCompanyNumber("12345678");
+        request.setPenaltyReference("PR123");
+
+        when(appealsService.delete(
+                request.getCompanyNumber(), request.getPenaltyReference()))
+                .thenReturn(true);
+
+        ResponseEntity<Void> response = appealController.deleteAppeal(request);
+
+        assertNull(response.getBody());
+        assertEquals(HttpStatus.NO_CONTENT, response.getStatusCode());
+        verify(appealsService, times(1))
+                .delete(request.getCompanyNumber(), request.getPenaltyReference());
+    }
+
+    @Test
+    void deleteAppealNotFound() {
+        DeleteAppealsRequest request = new DeleteAppealsRequest();
+        request.setCompanyNumber("12345678");
+        request.setPenaltyReference("PR123");
+
+        when(appealsService.delete(
+                request.getCompanyNumber(), request.getPenaltyReference()))
+                .thenReturn(false);
+
+        ResponseEntity<Void> response = appealController.deleteAppeal(request);
+
+        assertNull(response.getBody());
+        assertEquals(HttpStatus.NOT_FOUND, response.getStatusCode());
+        verify(appealsService, times(1))
+                .delete(request.getCompanyNumber(), request.getPenaltyReference());
+    }
+
+    @Test
+    void deleteAppealBadRequest() {
+        ResponseEntity<Void> response = appealController.deleteAppeal(null);
+
+        assertNull(response.getBody());
+        assertEquals(HttpStatus.BAD_REQUEST, response.getStatusCode());
+        verify(appealsService, times(0)).delete(anyString(), anyString());
+    }
+}
